Move newWalletPublicKeyHash to cmd package
We docode wallet public key hash from string in `cmd` pacakge and this
is the right place the `newWalletPublicKeyHash` function should be
placed.
